Add lazy import mechanism to the base package (#19838)
