+++ datadog_checks_base/changelog.d/19838.added
@@ -1,0 +1,1 @@
+Add lazy import mechanism
