+++ datadog_checks_base/datadog_checks/base/__init__.pyi
@@ -1,0 +1,31 @@
+# (C) Datadog, Inc. 2025-present
+# All rights reserved
+# Licensed under a 3-clause BSD style license (see LICENSE)
+from .__about__ import __version__
+from .agent import datadog_agent
+from .checks import AgentCheck
+from .checks.kube_leader import KubeLeaderElectionBaseCheck
+from .checks.openmetrics import OpenMetricsBaseCheck
+from .checks.openmetrics.v2.base import OpenMetricsBaseCheckV2
+from .checks.win import PDHBaseCheck
+from .checks.windows.perf_counters import PerfCountersBaseCheck
+from .config import is_affirmative
+from .errors import ConfigurationError
+from .utils.common import ensure_bytes, ensure_unicode, to_native_string, to_string
+
+__all__ = [
+    '__version__',
+    'AgentCheck',
+    'ConfigurationError',
+    'KubeLeaderElectionBaseCheck',
+    'OpenMetricsBaseCheck',
+    'OpenMetricsBaseCheckV2',
+    'PDHBaseCheck',
+    'PerfCountersBaseCheck',
+    'datadog_agent',
+    'ensure_bytes',
+    'ensure_unicode',
+    'is_affirmative',
+    'to_native_string',
+    'to_string',
+]
