+++ datadog_checks_base/datadog_checks/base/checks/__init__.pyi
@@ -1,0 +1,7 @@
+# (C) Datadog, Inc. 2025-present
+# All rights reserved
+# Licensed under a 3-clause BSD style license (see LICENSE)
+from .base import AgentCheck
+from .network import EventType, NetworkCheck, Status
+
+__all__ = ['AgentCheck', 'EventType', 'NetworkCheck', 'Status']
