+++ datadog_checks_base/datadog_checks/base/checks/kube_leader/__init__.py
@@ -1,9 +1,6 @@
 # (C) Datadog, Inc. 2018-present
 # All rights reserved
 # Licensed under a 3-clause BSD style license (see LICENSE)
+import lazy_loader
 
-from .base_check import KubeLeaderElectionBaseCheck
-from .mixins import KubeLeaderElectionMixin
-from .record import ElectionRecordAnnotation, ElectionRecordLease
-
-__all__ = ['KubeLeaderElectionMixin', 'ElectionRecordAnnotation', 'ElectionRecordLease', 'KubeLeaderElectionBaseCheck']
+__getattr__, __dir__, __all__ = lazy_loader.attach_stub(__name__, __file__)
