+++ datadog_checks_base/datadog_checks/base/checks/kube_leader/__init__.pyi
@@ -1,0 +1,8 @@
+# (C) Datadog, Inc. 2025-present
+# All rights reserved
+# Licensed under a 3-clause BSD style license (see LICENSE)
+from .base_check import KubeLeaderElectionBaseCheck
+from .mixins import KubeLeaderElectionMixin
+from .record import ElectionRecordAnnotation, ElectionRecordLease
+
+__all__ = ['ElectionRecordAnnotation', 'ElectionRecordLease', 'KubeLeaderElectionBaseCheck', 'KubeLeaderElectionMixin']
