+++ datadog_checks_base/datadog_checks/base/checks/kubelet_base/__init__.pyi
@@ -1,6 +1,6 @@
 # (C) Datadog, Inc. 2025-present
 # All rights reserved
 # Licensed under a 3-clause BSD style license (see LICENSE)
-import lazy_loader
+from .base import KubeletBase
 
-__getattr__, __dir__, __all__ = lazy_loader.attach_stub(__name__, __file__)
+__all__ = ['KubeletBase']
