+++ datadog_checks_base/datadog_checks/base/checks/openmetrics/v2/transformers/__init__.pyi
@@ -1,0 +1,26 @@
+# (C) Datadog, Inc. 2025-present
+# All rights reserved
+# Licensed under a 3-clause BSD style license (see LICENSE)
+from .counter import get_counter
+from .counter_gauge import get_counter_gauge
+from .gauge import get_gauge
+from .histogram import get_histogram
+from .metadata import get_metadata
+from .rate import get_rate
+from .service_check import get_service_check
+from .summary import get_summary
+from .temporal_percent import get_temporal_percent
+from .time_elapsed import get_time_elapsed
+
+__all__ = [
+    'get_counter',
+    'get_counter_gauge',
+    'get_gauge',
+    'get_histogram',
+    'get_metadata',
+    'get_rate',
+    'get_service_check',
+    'get_summary',
+    'get_temporal_percent',
+    'get_time_elapsed',
+]
