+++ datadog_checks_base/datadog_checks/base/checks/prometheus/__init__.py
@@ -1,9 +1,6 @@
 # (C) Datadog, Inc. 2018-present
 # All rights reserved
 # Licensed under a 3-clause BSD style license (see LICENSE)
+import lazy_loader
 
-from .base_check import GenericPrometheusCheck, PrometheusScraper
-from .mixins import PrometheusFormat, UnknownFormatError
-from .prometheus_base import PrometheusCheck
-
-__all__ = ['PrometheusFormat', 'UnknownFormatError', 'PrometheusCheck', 'GenericPrometheusCheck', 'PrometheusScraper']
+__getattr__, __dir__, __all__ = lazy_loader.attach_stub(__name__, __file__)
