+++ datadog_checks_base/datadog_checks/base/checks/prometheus/__init__.pyi
@@ -1,0 +1,8 @@
+# (C) Datadog, Inc. 2025-present
+# All rights reserved
+# Licensed under a 3-clause BSD style license (see LICENSE)
+from .base_check import GenericPrometheusCheck, PrometheusScraper
+from .mixins import PrometheusFormat, UnknownFormatError
+from .prometheus_base import PrometheusCheck
+
+__all__ = ['GenericPrometheusCheck', 'PrometheusCheck', 'PrometheusFormat', 'PrometheusScraper', 'UnknownFormatError']
