+++ datadog_checks_base/datadog_checks/base/checks/win/wmi/__init__.pyi
@@ -1,0 +1,32 @@
+# (C) Datadog, Inc. 2025-present
+# All rights reserved
+# Licensed under a 3-clause BSD style license (see LICENSE)
+from .base import (
+    InvalidWMIQuery,
+    MissingTagBy,
+    TagQuery,
+    TagQueryUniquenessFailure,
+    WinWMICheck,
+    WMIFilter,
+    WMIMetric,
+    WMIObject,
+    WMIProperties,
+    WMISampler,
+    from_time,
+    to_time,
+)
+
+__all__ = [
+    'InvalidWMIQuery',
+    'MissingTagBy',
+    'TagQuery',
+    'TagQueryUniquenessFailure',
+    'WinWMICheck',
+    'WMIFilter',
+    'WMIMetric',
+    'WMIObject',
+    'WMIProperties',
+    'WMISampler',
+    'from_time',
+    'to_time',
+]
