+++ datadog_checks_base/datadog_checks/base/utils/db/__init__.py
@@ -1,5 +1,6 @@
 # (C) Datadog, Inc. 2019-present
 # All rights reserved
 # Licensed under a 3-clause BSD style license (see LICENSE)
-from .core import QueryExecutor, QueryManager
-from .query import Query
+import lazy_loader
+
+__getattr__, __dir__, __all__ = lazy_loader.attach_stub(__name__, __file__)
