+++ datadog_checks_base/datadog_checks/base/utils/db/__init__.pyi
@@ -1,0 +1,7 @@
+# (C) Datadog, Inc. 2025-present
+# All rights reserved
+# Licensed under a 3-clause BSD style license (see LICENSE)
+from .core import QueryExecutor, QueryManager
+from .query import Query
+
+__all__ = ['Query', 'QueryExecutor', 'QueryManager']
