+++ datadog_checks_base/datadog_checks/base/utils/metadata/__init__.py
@@ -1,4 +1,6 @@
 # (C) Datadog, Inc. 2019-present
 # All rights reserved
 # Licensed under a 3-clause BSD style license (see LICENSE)
-from .core import MetadataManager
+import lazy_loader
+
+__getattr__, __dir__, __all__ = lazy_loader.attach_stub(__name__, __file__)
