+++ datadog_checks_base/datadog_checks/base/utils/metadata/__init__.pyi
@@ -1,6 +1,6 @@
 # (C) Datadog, Inc. 2025-present
 # All rights reserved
 # Licensed under a 3-clause BSD style license (see LICENSE)
-import lazy_loader
+from .core import MetadataManager
 
-__getattr__, __dir__, __all__ = lazy_loader.attach_stub(__name__, __file__)
+__all__ = ['MetadataManager']
