+++ datadog_checks_base/datadog_checks/base/utils/prometheus/__init__.py
@@ -1,5 +1,6 @@
 # (C) Datadog, Inc. 2016-present
 # All rights reserved
 # Licensed under Simplified BSD License (see LICENSE)
+import lazy_loader
 
-from .functions import parse_metric_family  # noqa: F401
+__getattr__, __dir__, __all__ = lazy_loader.attach_stub(__name__, __file__)
