+++ datadog_checks_base/tests/base/checks/windows/conftest.py
@@ -1,6 +1,8 @@
 # (C) Datadog, Inc. 2025-present
 # All rights reserved
 # Licensed under a 3-clause BSD style license (see LICENSE)
-import lazy_loader
+import sys
 
-__getattr__, __dir__, __all__ = lazy_loader.attach_stub(__name__, __file__)
+
+def pytest_ignore_collect(path, config):
+    return sys.platform != 'win32'
